refactor: Restructure project with src-layout to fix imports
