+++ src/i_drill_project.egg-info/PKG-INFO
@@ -1,0 +1,3 @@
+Metadata-Version: 2.4
+Name: i_drill_project
+Version: 0.1.0
