+++ tests/test_data_reconciler.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-from Scripts.data_reconciler import DataReconciler
+from src.Scripts.data_reconciler import DataReconciler
 
 import unittest
